bc포맷 관련 pitch계산 로직 수정. Cam클래스에 Buffer하나만 두고 offse을 통해 나눠서 사용.

+++ Renderer/LightBuffer12.h
@@ -80,8 +80,8 @@ namespace wilson
 		std::vector<PointLight12*>m_pPointLights;
 		std::vector<SpotLight12*>m_pSpotLights;
 
-		UINT* m_pLightPropertyCbBegin;
-		UINT* m_pDirLitMatricesCbBegin;
-		UINT* m_pSpotLitMatricesCbBegin;
+		UINT8* m_pLightPropertyCbBegin;
+		UINT8* m_pDirLitMatricesCbBegin;
+		UINT8* m_pSpotLitMatricesCbBegin;
 	};
 }
